sistema_bellafly/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-jazzmin==3.0.1
idna==3.10
pillow==11.2.1
requests==2.32.4
sqlparse==0.5.3
types-requests==2.32.4.20250611
urllib3==2.5.0
gunicorn==21.2.0
psycopg2-binary